pyproject.toml
[project]
name = "portfolio-evolution"
version = "0.1.0"
description = "Daily Portfolio Evolution & Pipeline Simulation Engine"
requires-python = ">=3.11"
dependencies = [
    "polars>=1.0",
    "pandas>=2.0",
    "numpy>=1.26",
    "pydantic>=2.0",
    "scipy>=1.12",
    "pyyaml>=6.0",
    "typer>=0.12",
    "rich>=13.0",
    "duckdb>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "apscheduler>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "hypothesis>=6.0",
]
notebooks = [
    "plotly>=5.0",
    "jupyter>=1.0",
]

[project.scripts]
portfolio-evolution = "portfolio_evolution.main:app"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
